Initial commit of new Docker based set up.
This project is basically put in stasis mode by converting it to run in
two Docker containers. The Omeka file structure and the database file
are accessible in the file structure because of the size of the data in
the Omeka installation (over 5GB) and to keep the user account
information out of an image for the database.

Files Changed:
- README
  - Updates the README for information on setting up development and
production environments
- .gitignore
  - ignores environment variables file, the omeka folder, and the
  database folder
- Dockerfile
  - Adds a file for creating the image used for Omeka. Image created
  this way to include ImageMagick and Sendmail
- config/php.ini
  - Changes the default php.ini to allow for larger file uploads
- docker-compose.yml
  - File for starting the Omeka and MySQL containers. This version also
  has settings for utilizing Traefik.
- upgrade-omeka
  - file removed as not needed for upgrading Omeka